Fix: Convert int to long for HP and damage calculations

+++ Source By Mr Blue/src/nro/models/Bot/BotAttackplayer.java
@@ -222,7 +222,7 @@ public class BotAttackplayer extends Bot {
             return;
         }
 
-        int damePST = (int) (dame * percentPST / 100L);
+        long damePST = dame * percentPST / 100L;
         if (damePST >= attacker.nPoint.hp) {
             damePST = attacker.nPoint.hp - 1;
         }
@@ -232,7 +232,7 @@ public class BotAttackplayer extends Bot {
             msg = new Message(56);
             msg.writer().writeInt((int) attacker.id);
             damePST = attacker.injured(attacker, damePST, true, false);
-            msg.writer().writeInt(attacker.nPoint.hp);
+            msg.writer().writeLong(attacker.nPoint.hp);
             msg.writer().writeInt(damePST);
             msg.writer().writeBoolean(false);
             msg.writer().writeByte(36);
