+++ Source By Mr Blue/src/nro/models/boss/doanh_trai/NinjaClone.java
@@ -20,13 +20,13 @@ public class NinjaClone extends Boss {
 
     private Boss boss;
 
-    public NinjaClone(Zone zone, Boss boss, int dame, int hp, int id) throws Exception {
+    public NinjaClone(Zone zone, Boss boss, long dame, long hp, int id) throws Exception {
         super(PHOBANDT, id, new BossData(
                 "Ninja Áo Tím", //name
                 ConstPlayer.TRAI_DAT, //gender
                 new short[]{123, 124, 125, -1, -1, -1}, //outfit {head, body, leg, bag, aura, eff}
                 ((dame)), //dame
-                new int[]{((hp))}, //hp
+                new int[]{((int)hp)}, //hp
                 new int[]{54}, //map join
                 new int[][]{
                     {Skill.DEMON, 3, 1}, {Skill.DEMON, 6, 2}, {Skill.DRAGON, 7, 3}, {Skill.DRAGON, 1, 4}, {Skill.GALICK, 5, 5},
